[ci] Updating lmi-dist ci tests for rubikon-engine (#1651)
